Simplify and improve GIC-less Cortex-R4 port. Add final tests into RZ/T demo.

+++ FreeRTOS/Demo/Common/Minimal/TaskNotify.c
@@ -556,6 +556,6 @@ const uint32_t ulMultiplier = 0x015a4e35UL, ulIncrement = 1UL;
 
 	/* Utility function to generate a pseudo random number. */
 	ulNextRand = ( ulMultiplier * ulNextRand ) + ulIncrement;
-	return( ( int ) ( ulNextRand >> 16UL ) & 0x7fffUL );
+	return( ( ulNextRand >> 16UL ) & 0x7fffUL );
 }
 /*-----------------------------------------------------------*/
